+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Breaking changes in **bold**.
 #### v0.9.1 – Jul 2020
 
 * **Fix: throttle operator behaved more like debounce, it was not emitting until the parent stopped emitting for `intervalMillis`. It behaves as advertised now. ([#34](https://github.com/raquo/Airstream/issues/34))**
-
+* New: `Observer.combine(observer1, observer2, ...)` factory
+* New: `observer.contracollect { case Foo => bar }` operator
 
 #### v0.9.0 – Mar 2020